requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==2.0.4
djangorestframework==3.10.3
Flask==0.10.1
get==2019.4.13
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.2
numpy==1.17.4
post==2019.4.13
public==2019.4.13
pycurl==7.43.0.3
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.18.4
urllib3==1.22
Werkzeug==0.16.0
